Move side effect test cases out of marker tests (to enalbe non-shallow analysis)
