+++ crates/paralegal-flow/tests/marker_tests.rs
@@ -317,208 +317,3 @@ fn async_fn_marker() {
         assert!(source.flows_to_data(&sink));
     });
 }
-
-#[test]
-fn side_effect_tcp() {
-    inline_test! {
-        use std::io::prelude::*;
-        use std::net::TcpStream;
-
-        fn main() -> std::io::Result<()> {
-            let mut stream = TcpStream::connect("127.0.0.1:34254")?;
-
-            stream.write(&[1])?;
-            stream.read(&mut [0; 128])?;
-            Ok(())
-        }
-    }
-    .with_extra_args(["--side-effect-markers".to_string()])
-    .check_ctrl(|ctrl| {
-        let auto_markers = AutoMarkers::new();
-        let defined = ctrl.markers();
-        let auto = auto_markers.all();
-        let contained = dbg!(auto
-            .iter()
-            .filter(|m| defined.contains(m))
-            .collect::<Vec<_>>());
-        assert!(!contained.is_empty());
-    });
-}
-
-#[test]
-fn side_effect_pure() {
-    inline_test! {
-        fn main() -> std::io::Result<()> {
-            let x = 2;
-            let y = 3 + x;
-            Ok(())
-        }
-    }
-    .with_extra_args(["--side-effect-markers".to_string()])
-    .check_ctrl(|ctrl| {
-        let auto_markers = AutoMarkers::new();
-        let defined = ctrl.markers();
-        let auto = auto_markers.all();
-        let contained = dbg!(auto
-            .iter()
-            .filter(|m| defined.contains(m))
-            .collect::<Vec<_>>());
-        assert!(contained.is_empty());
-    });
-}
-
-#[test]
-fn side_effect_extern() {
-    inline_test! {
-        extern "C" {
-            fn plus(a: i32, b: i32) -> i32;
-        }
-
-        fn main() -> std::io::Result<()> {
-            let x = 2;
-            let y = plus(3, x);
-            Ok(())
-        }
-    }
-    .with_extra_args(["--side-effect-markers".to_string()])
-    .check_ctrl(|ctrl| {
-        let auto_markers = AutoMarkers::new();
-        let defined = dbg!(ctrl.markers());
-        let auto = auto_markers.all();
-        for i in ctrl.functions("plus") {
-            dbg!(ctrl.info_for(i.ident));
-        }
-        let contained = dbg!(auto
-            .iter()
-            .filter(|m| defined.contains(m))
-            .collect::<Vec<_>>());
-        assert!(!contained.is_empty());
-    });
-}
-
-#[test]
-fn side_effect_extern_flow() {
-    inline_test! {
-        extern "C" {
-            fn plus(a: i32, b: i32) -> i32;
-        }
-
-        #[paralegal_flow::marker(source, return)]
-        fn source() -> i32 {
-            42
-        }
-
-        #[paralegal_flow::marker(source2, return)]
-        fn source2() -> i32 {
-            43
-        }
-
-        fn main() -> std::io::Result<()> {
-            let x = source2();
-            let z = source();
-            let y = plus(z, 3);
-            let result = y + x;
-            Ok(())
-        }
-    }
-    .with_extra_args(["--side-effect-markers".to_string()])
-    .check_ctrl(|ctrl| {
-        let auto_markers = AutoMarkers::new();
-        let defined = dbg!(ctrl.markers());
-        let auto = auto_markers.all();
-        let contained = dbg!(auto
-            .iter()
-            .filter(|m| defined.contains(m))
-            .collect::<Vec<_>>());
-        assert!(!contained.is_empty());
-
-        let source1 = ctrl.marked("source");
-        let source2 = ctrl.marked("source2");
-        let side_effecting: NodeRefs = auto_markers
-            .all()
-            .iter()
-            .flat_map(|m| ctrl.marked(*m))
-            .collect();
-        assert!(!source1.is_empty());
-        assert!(!source2.is_empty());
-        assert!(!side_effecting.is_empty());
-
-        assert!(source1.flows_to_data(&side_effecting));
-        assert!(!source2.flows_to_data(&side_effecting));
-    });
-}
-
-fn tcp_flow_check(ctrl: CtrlRef<'_>) {
-    let auto_markers = AutoMarkers::new();
-    let defined = dbg!(ctrl.markers());
-    let auto = auto_markers.all();
-    let contained = dbg!(auto
-        .iter()
-        .filter(|m| defined.contains(m))
-        .collect::<Vec<_>>());
-    assert!(!contained.is_empty());
-
-    let source1 = ctrl.marked("source");
-    let source2 = ctrl.marked("source2");
-    let side_effecting: NodeRefs = auto_markers
-        .all()
-        .iter()
-        .flat_map(|m| ctrl.marked(*m))
-        .collect();
-    assert!(!source1.is_empty());
-    assert!(!source2.is_empty());
-    assert!(!side_effecting.is_empty());
-
-    assert!(source1.flows_to_data(&side_effecting));
-    assert!(!source2.flows_to_data(&side_effecting));
-}
-
-define_test!(side_effect_tcp_flow_in_separate_crate: ctrl -> {
-    tcp_flow_check(ctrl);
-});
-
-#[test]
-fn side_effect_tcp_flow() {
-    inline_test! {
-        use std::io::prelude::*;
-        use std::net::TcpStream;
-
-        #[paralegal_flow::marker(source, return)]
-        fn source() -> u8 {
-            42
-        }
-
-        #[paralegal_flow::marker(source2, return)]
-        fn source2() -> u8 {
-            43
-        }
-
-        fn main() -> std::io::Result<()> {
-            let mut stream = TcpStream::connect("127.0.0.1:34254")?;
-            let y = source2();
-            stream.write(&[source()])?;
-            let mut buf = [0; 128];
-            stream.read(&mut buf)?;
-            let _ = y + buf[0];
-            Ok(())
-        }
-
-    }
-    .with_extra_args(["--side-effect-markers".to_string()])
-    .check_ctrl(|ctrl| {
-        tcp_flow_check(ctrl);
-    });
-}
-
-#[test]
-fn side_effect_vec() {
-    inline_test! {
-        fn main() {
-            let mut v = vec![0];
-            v.push(1);
-            v.pop();
-        }
-    }
-    .with_extra_args(["--side-effect-markers"])
-    .check_ctrl(|ctrl| ctrl.assert_purity(true));
-}
