+++ crates/paralegal-flow/tests/purity-tests.rs
@@ -1,3 +1,6 @@
 #![feature(rustc_private)]
 
+#[macro_use]
+extern crate lazy_static;
+
 mod purity;
