+++ crates/paralegal-flow/tests/purity/test-crate/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "purity-tests"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+paralegal = { path = "../../../../paralegal" }
